dom/workers/test/indexedDB/test_concurrent_access.html

<!--
  Any copyright is dedicated to the Public Domain.
  http://creativecommons.org/publicdomain/zero/1.0/
-->
<html>
<head>
  <title>Indexed Database Property Test</title>

  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>

  <script type="text/javascript;version=1.7">
  const indexedDBSyncPref = "dom.indexedDBSync.enabled";
  const numberOfWorkers = 10;

  const data = {
    name: window.location.pathname,
    objectStoreName: "Objects",
    objectStoreData: { value: { name: "Marvin" }, key: 1 },
    indexData: { name: "nameIndex", key: "name" }
  };

  // Create db using async API.
  let request = window.indexedDB.open(data.name, 1);

  request.onupgradeneeded = function(event) {
    info("Initialization");
    let db = event.target.result;
    var objectStore = db.createObjectStore(data.objectStoreName, { autoIncrement: 0 });
    objectStore.add(data.objectStoreData.value, data.objectStoreData.key);
    objectStore.createIndex(data.indexData.name, data.indexData.key);
  };

  request.onsuccess = function(event) {
    info("Initialization succeeded");

    SpecialPowers.pushPrefEnv({ set: [[indexedDBSyncPref, true]] }, function() {
      // Let's run several concurrent workers at once.
      let counter = 0;
      for (let i = 0; i < numberOfWorkers; i++) {
        let worker = new Worker("concurrent_access_worker.js");
        worker.onmessage = function(event) {
          let result = event.data;
          if(result == undefined) {
            counter++;
            info(counter + ". worker finished");
            if (counter == numberOfWorkers) {
              info("Test successfully completed");
              SimpleTest.finish();
            }
          }
          else {
            if (result.type == "ok") {
              ok(result.condition, result.name, result.diag);
            }
          }
        };
        worker.postMessage(data);
      }
    });
  };

  SimpleTest.waitForExplicitFinish();
  </script>

</head>

<body></body>

</html>
